Fix 11 pre-existing clippy lints blocking repair branches
The quality gate (cargo clippy -- -D warnings) was failing on main,
causing all prole branches to fail the gate and enter repairing status.

- Add #[allow(clippy::too_many_arguments)] to 6 Bevy system fns
  (idiomatic — ECS systems need many query params)
- Replace for (_, v) in &map with map.values()
- Replace repeat().take() with repeat_n()
- Replace manual range check with Range::contains
- Remove unnecessary u32 cast
- Replace iter().map(|(_, v)|) with values().map(|v|)

+++ src/calendar.rs
@@ -86,7 +86,7 @@ pub fn first_working_day_of_month(year: i32, month: u32, config: &CalendarConfig
         config.non_working_dates.iter().copied().collect();
     let mut day = NaiveDate::from_ymd_opt(year, month, 1)?;
     loop {
-        if day.month() != month as u32 {
+        if day.month() != month {
             return None;
         }
         if is_working_day(day, config.working_days_per_week, &non_working) {
